fix: resolve ESLint config error and SSR window crash in build
