+++ .eslintrc.json
@@ -9,10 +9,12 @@
   "rules": {
     "react/no-unescaped-entities": "off",
     "@next/next/no-page-custom-font": "off",
-    "react/display-name": "off" ,
-    "react/no-unknown-property": ["error", { "ignore": ["suppressHydrationWarning"] }],
-    "no-unused-vars": ["warn", { "suppressWarnings": ["unused-imports"] }],
-    "react/jsx-key": "off"
-
+    "react/display-name": "off",
+    "react/no-unknown-property": ["warn", { "ignore": ["suppressHydrationWarning", "sx", "maxRows", "minRows", "container", "item", "justifyContent", "alignItems", "http-equiv", "crossorigin", "clip-path"] }],
+    "no-unused-vars": ["warn", { "varsIgnorePattern": "^_", "argsIgnorePattern": "^_" }],
+    "react/jsx-key": "off",
+    "react-hooks/rules-of-hooks": "warn",
+    "react/jsx-no-undef": "warn",
+    "react/no-children-prop": "warn"
   }
 }
